Be a little less noisy on github issues mentioned outside of meetings.
In particular, this stops giving the combined message:
  I can't set a github URL because you haven't started a topic.  Also, Because I don't want to spam github issues unnecessarily, I won't comment in that github issue unless you write "Github: <issue-url> | none" (or "Github issue: ..."/"Github topic: ...").
because if both of those conditions fail, it's better to just keep
quiet.

But it still continues giving combined messages when the "Github:"
(etc.) syntax is used.

This also avoids the incorrect casing in that particular combined
message that @tabatkins pointed out on #css IRC.