Add the ability to comment (though not remove labels) from PRs.
This generalizes the commenting (but not label removal) code that
existed for issues to also work on pulls.  hubcaps doesn't appear to
support label manipulation on pull requests, so that will have to wait.

I'm also unhappy about the amount of code I could share, but the
alternatives seemed worse...

Fixes #19.

(Tested in a testing configuration to produce a comment on #23, to make
sure the real github interaction works.)